+++ packages/expo-router/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Allow the file extension to be specified for `renderRouter`'s filepaths ([#26510](https://github.com/expo/expo/pull/26510) by [@marklawlor](https://github.com/marklawlor))
 - Add `useNavigationContainerRef` to access the root NavigationContainer ref. ([#26529](https://github.com/expo/expo/pull/26529) by [@EvanBacon](https://github.com/EvanBacon))
+- Allow `renderRouter()` to accept an array of strings to quickly mock multiple empty components.
 
 ### 🐛 Bug fixes